Adds Game.java.
Adds a new file, Game.java. It will be the main fiel of the game. #
Please enter the commit message for your changes. Lines starting # with
'#' will be ignored, and an empty message aborts the commit. # On branch
maingame # Changes to be committed: # (use "git reset HEAD <file>..." to
unstage) # # new file: Game.java #